pyproject.toml
[project]
name = "venture-vision-fastapi-backend"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Paras", email = "parasbhardava26@gmail.com" }
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi>=0.115.8,<0.116.0",
    "uvicorn[standard]>=0.34.0,<0.35.0",
    "gunicorn>=23.0.0,<24.0.0",
    "uvloop>=0.21.0,<0.22.0",
    "sqlalchemy[asyncio]>=2.0.41,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "yarl>=1.9.0,<2.0.0",
    "pydantic>=2.11.5,<3.0.0",
    "loguru>=0.7.3,<1.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "psycopg[c]>=3.2.4,<4.0.0",
    "alembic>=1.16.1,<2.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "sqlmodel>=0.0.24,<0.1.0",
    "google-adk>=1.13.0",
    "streamlit>=1.49.1",
    "pandas>=2.3.2",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0,<5.0.0",
    "pylint>=3.3.7,<4.0.0",
    "mypy>=1.16.0,<2.0.0",
    "ruff>=0.11.13,<1.0.0",
    "types-pyyaml>=6.0.12.20250516,<7.0.0",
    "pytest>=8.4.0,<9.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

